AI-Service/requirements.txt
flask
flask-cors
openai
python-dotenv
sqlalchemy
pymysql
httpx
psutil
transformers
torch
sentence-transformers
faiss-cpu
numpy
scikit-learn
chromadb
PyJWT